requirements.txt
pyspark
pyhive
nibabel
nilearn
scikit-learn
pandas
numpy
thrift
thrift_sasl
hdfs
pyyaml
xgboost